Modify removeCommand for exercise
