+++ src/main/java/seedu/duke/command/RemoveCommand.java
@@ -4,7 +4,6 @@ import seedu.duke.Parser;
 import seedu.duke.Ui;
 import seedu.duke.biometrics.Biometrics;
 import seedu.duke.exception.IllegalValueException;
-import seedu.duke.exercise.Exercise;
 import seedu.duke.exercise.ExerciseList;
 import seedu.duke.food.FoodList;
 import seedu.duke.storage.Storage;
@@ -14,6 +13,7 @@ public class RemoveCommand extends Command {
     private FoodList foodList;
     private String arguments;
     public static final String INVALID_REMOVE_FOOD_INPUT = "Invalid remove food input";
+    private ExerciseList exerciseList;
 
     public RemoveCommand(String arguments) {
         this.arguments = arguments;
@@ -27,8 +27,29 @@ public class RemoveCommand extends Command {
         case ("food"):
             removeFood(argumentList);
             break;
+        case ("exercise"):
+            removeExercise(argumentList);
+            break;
         default:
             handleInvalidRemoveType();
+        }
+    }
+
+    //@@junhaoliu2468
+    private void removeExercise(String[] argumentList) {
+        try {
+            if (argumentList.length != 2) {
+                throw new IllegalValueException("INVALID_NUMBER_INPUT");
+            }
+            int index = Integer.parseInt(argumentList[1]);
+            if (index > exerciseList.getCurrentExerciseList().size() || index < 1) {
+                throw new IllegalValueException("INVALID_INDEX_INPUT");
+            }
+            String exerciseName = exerciseList.getCurrentExercise(index - 1).getExerciseName();
+            exerciseList.removeCurrentExercise(index - 1);
+            ui.output(exerciseName + " has been removed successfully!");
+        } catch (IllegalValueException | NumberFormatException e) {
+            ui.output(e.getMessage());
         }
     }
 
@@ -60,5 +81,6 @@ public class RemoveCommand extends Command {
     public void setData(Ui ui, Storage storage, Biometrics biometrics, ExerciseList exerciseList, FoodList foodList) {
         this.ui = ui;
         this.foodList = foodList;
+        this.exerciseList = exerciseList;
     }
 }
